extras/chop/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chop"
version = "0.1.0"
description = "Unix-philosophy image manipulation CLI with JSON piping for terminal graphics"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "charpx developers"}
]
keywords = ["terminal", "image", "manipulation", "pipeline", "unix", "braille", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Terminals",
]
dependencies = [
    "numpy>=1.20",
    "pillow>=9.0",
    "charpx>=0.1.0",
]

[project.scripts]
chop = "chop:main"

[project.urls]
Homepage = "https://github.com/anthropics/charpx"
Repository = "https://github.com/anthropics/charpx"
Documentation = "https://github.com/anthropics/charpx/tree/main/extras/chop"

[tool.setuptools.packages.find]
where = ["."]
